configs/manager.py: Implement "setn()" to set an option value
The "setn()" method is a corresponding operation with the "getn()",
which set the value of a config option to the given value.

This function will be used in the Web UI to interact with the WebSocket
communications.

Also add the "merge()" method which simply merge the input
configurations without any validation.