webui: Replace jQuery.when() with jQuery.Deferred's .then()
Note that jQuery.when() can NOT promise the ordering of the passed
Deferred's.  Therefore, chain multiple AJAX requests using ".then()" and
".done()" to keep them ordered.

This fixes the configurations loading problem on IE (v11), which calls
the "when()" arguments in different orders as Firefox and Chromium, thus
leads to the wrong results.