products.py: Add parameter "load" to __init__()
The "load" parameter explicitly controls whether to load the specified
manifest file.

Update "foregrounds.py" to use this new "load" parameter, which prevent
the manifest file been loaded.  Otherwise, if the output manifest file
already exists, it will be loaded to Products, which may cause
inconsistencies with the manifest of newly simulated products.