clusters/emission: Rewrite synchrotron kernel function
Add asymptotic functions to calculate the values beyond the
interpolation bounds (e.g., <1e-3 and >10), otherwise, the calculated
synchrotron emissivity is overestimated at the higher frequencies.

By rewrite this synchrotron kernel function, the calculated results is
consistent with the theoretical/analytical results, e.g., the
synchrotron radiation of a population of electrons of power-law index n
is also a power-law with index (n-1)/2.