clusters/halo: Rewrite turbulence acceleration & diffusion coefficient!
The previous calculation of the turbulence acceleration efficiency has
significant problems, which is too low, i.e., the acceleration timescale is
too large (~10 Gyr).  However, it is reported that turbulent acceleration has
a timescale ~100 Myr (0.1 Gyr).  I believe this problem is due to the
over-simplification to the formula proposed by [cassano2005].

Rewrite the turbulence acceleration and diffusion coefficient calculations,
adopting the method proposed by [brunetti2016].  To this end, two new options
"f_lturb" and "f_acc" are introduced to tune the results.