clusters/halo: Adopt new method for _velocity_turb()
Adopt a new method to calculate the turbulence velocity dispersion.
First, the kinetic energy released by the merger is estimated as the
work done by the in-falling sub-cluster crossing the main cluster.
Second, a fraction of the kinetic energy is assumed to be transferred
into the turbulence.  The turbulence mass is calculated as the gas
mass enclosed within the turbulence region.  Finally, the turbulence
velocity dispersion is derived.